requirements.txt
# FastAPI 与运行
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
jinja2>=3.1.0
python-multipart>=0.0.6
itsdangerous>=2.1.0
python-dotenv>=1.0.0

# 数据与密码
sqlalchemy>=2.0.0
bcrypt>=4.0.0,<5.0.0

# DOCX 与 GitHub
python-docx>=1.1.0
PyGithub>=2.1.0
